+++ lib/src/api/objects/authentication_code_type_sms.dart
@@ -21,7 +21,7 @@ class AuthenticationCodeTypeSms extends AuthenticationCodeType {
     }
 
     return AuthenticationCodeTypeSms(
-      length: json['length'],
+      length: json['length'] as int,
     );
   }
