+++ lib/src/api/objects/bot_command_scope_chat.dart
@@ -20,7 +20,7 @@ class BotCommandScopeChat extends BotCommandScope {
     }
 
     return BotCommandScopeChat(
-      chatId: json['chat_id'],
+      chatId: json['chat_id'] as int,
     );
   }
