+++ lib/src/api/objects/can_transfer_ownership_result_session_too_fresh.dart
@@ -23,7 +23,7 @@ class CanTransferOwnershipResultSessionTooFresh
     }
 
     return CanTransferOwnershipResultSessionTooFresh(
-      retryAfter: json['retry_after'],
+      retryAfter: json['retry_after'] as int,
     );
   }
