+++ lib/src/api/objects/chat_event_slow_mode_delay_changed.dart
@@ -24,8 +24,8 @@ class ChatEventSlowModeDelayChanged extends ChatEventAction {
     }
 
     return ChatEventSlowModeDelayChanged(
-      oldSlowModeDelay: json['old_slow_mode_delay'],
-      newSlowModeDelay: json['new_slow_mode_delay'],
+      oldSlowModeDelay: json['old_slow_mode_delay'] as int,
+      newSlowModeDelay: json['new_slow_mode_delay'] as int,
     );
   }
