+++ lib/src/api/objects/chat_members_filter_mention.dart
@@ -21,7 +21,7 @@ class ChatMembersFilterMention extends ChatMembersFilter {
     }
 
     return ChatMembersFilterMention(
-      messageThreadId: json['message_thread_id'],
+      messageThreadId: json['message_thread_id'] as int,
     );
   }
