+++ lib/src/api/objects/chat_type_secret.dart
@@ -24,8 +24,8 @@ class ChatTypeSecret extends ChatType {
     }
 
     return ChatTypeSecret(
-      secretChatId: json['secret_chat_id'],
-      userId: json['user_id'],
+      secretChatId: json['secret_chat_id'] as int,
+      userId: json['user_id'] as int,
     );
   }
