+++ lib/src/api/objects/email_address_reset_state_pending.dart
@@ -23,7 +23,7 @@ class EmailAddressResetStatePending extends EmailAddressResetState {
     }
 
     return EmailAddressResetStatePending(
-      resetIn: json['reset_in'],
+      resetIn: json['reset_in'] as int,
     );
   }
