+++ lib/src/api/objects/passport_element_error_source_file.dart
@@ -21,7 +21,7 @@ class PassportElementErrorSourceFile extends PassportElementErrorSource {
     }
 
     return PassportElementErrorSourceFile(
-      fileIndex: json['file_index'],
+      fileIndex: json['file_index'] as int,
     );
   }
