+++ lib/src/api/objects/passport_element_error_source_translation_file.dart
@@ -23,7 +23,7 @@ class PassportElementErrorSourceTranslationFile
     }
 
     return PassportElementErrorSourceTranslationFile(
-      fileIndex: json['file_index'],
+      fileIndex: json['file_index'] as int,
     );
   }
