+++ lib/src/api/objects/storage_statistics_fast.dart
@@ -37,11 +37,11 @@ class StorageStatisticsFast extends TdObject {
     }
 
     return StorageStatisticsFast(
-      filesSize: json['files_size'],
-      fileCount: json['file_count'],
-      databaseSize: json['database_size'],
-      languagePackDatabaseSize: json['language_pack_database_size'],
-      logSize: json['log_size'],
+      filesSize: json['files_size'] as int,
+      fileCount: json['file_count'] as int,
+      databaseSize: json['database_size'] as int,
+      languagePackDatabaseSize: json['language_pack_database_size'] as int,
+      logSize: json['log_size'] as int,
     );
   }
