+++ lib/src/api/objects/update_chat_message_auto_delete_time.dart
@@ -25,8 +25,8 @@ class UpdateChatMessageAutoDeleteTime extends Update {
     }
 
     return UpdateChatMessageAutoDeleteTime(
-      chatId: json['chat_id'],
-      messageAutoDeleteTime: json['message_auto_delete_time'],
+      chatId: json['chat_id'] as int,
+      messageAutoDeleteTime: json['message_auto_delete_time'] as int,
     );
   }
